Add llvm (version 11.0) to the project
In this commit the project only compiles with llvm successfuly.

To prove it, I have written a test that uses llvm to compile the expression
return 555 + 111, and checked that the result assembly contains the
assembly: "movl $666, eax".

To be able to compile the project, you need to extract the
llvm-11.0.src.tar.xz inside src/llvm and compile whatever is inside it.

In the future I should probably make this step automatic when first
trying to compile panther.