backend: llvm: Not rely on llvm types
LLVM types are not easy to work with.

For example when casting, I only know if the type is an integer, without
knowing if it's unsigned / signed and so on...

To handle it better in the backend, I introduced a new type that holds
both the llvm value and the type I created for the project.

In addition, I made all unknown casts bitcasts, in the future I want to
add better casting support, but for now this is fine.