backend: llvm: Add unconditional jump opcode
By implementing this opcode, I noticed there was a bug when building the
llvm IR.

In LLVM's docs it says that all basic blocks require a terminator, and
preferably a single terminator because then there would be unreachable /
higher precedence code.

So I rewrote the buildJump logic to create a basic block, put it inside
a struct, and access it later when we get to the jump location opcode.