frontend: C style blocks
Instead of defining blocks using `:` and indentation like in python /
nim, define them using curly brackets and semicolons line in C / Rust.

This also means that I changed comments from `#` to `//`.

The reason I wanted this change is because when I introduce lambda
function, I want them to be able to have more than a single expression.

In python when defining a lambda, for example: `lambda: None`, there can
only be one expression inside the lambda block, and the reason for that
is that it's impossible using indentation and new lines inside the lambda
without it looking like a totally new expression.