backend: llvm: Prefer small code than fast code
Minize size by not letting faster but larger code to be ever picked over
small code.

The way I did this is by setting the opt level to the lowest (0) and the
size level to highest (2).

Previously, I thought combining the two would be the best of both
worlds, but apperently after digging enough in llvm's rabbit hole, I
found out that this is not the case.

If you want the smallest code, you need to set it's level to be higher
than the fast code, as llvm chooses either fast or small code, not doing
both.